+++ assets/src/scripts/components/hydrograph/drawing-data.js
@@ -5,7 +5,6 @@ import {DateTime} from 'luxon';
 import {createSelector} from 'reselect';
 import {format} from 'd3-format';
 
-import config from '../../config';
 import {getCurrentVariableMedianStatistics} from '../../selectors/median-statistics-selector';
 import {getVariables, getCurrentMethodID, getTimeSeries, getCurrentVariableTimeSeries, getTimeSeriesForTsKey,
     getTsRequestKey, getRequestTimeRange} from '../../selectors/time-series-selector';
@@ -291,7 +290,7 @@ export const lineSegmentsSelector = memoize((tsKey, period) => createSelector(
 
             for (let pt of points) {
                 // Classes to put on the line with this point.
-                let lineClasses = getLineClasses(pt, !config.MULTIPLE_TIME_SERIES_METADATA_SELECTOR_ENABLED || !currentMethodID || currentMethodID === parseInt(methodID));
+                let lineClasses = getLineClasses(pt, !currentMethodID || currentMethodID === parseInt(methodID));
 
                 // If this is a non-masked data point, split lines if the gap
                 // from the period point exceeds MAX_LINE_POINT_GAP.